generalize the way screens are skipped when installing desktop
Also skip identity screen when installing neither server nor desktop
(i.e. core).
